requirements.txt
python-barcode == 0.14